Add CSS styling to change message
The main reason for the style is to fix displaying of long messages. The
long messages are wrapped and a scrollbar is added if needed to prevent
the message from taking too much space.

Also make the message more visible by using the same style as the
message fo the day.

Change-Id: Iaab66afa85b00bd901ae1d0c63398af6e667c7a0